Turn Purchases.framework into a cross-platform framework
This is done by providing a value for the `SUPPORTED_PLATFORMS` build setting that encompasses both the macOS and iOS platforms. Then, the build settings for the macOS and iOS frameworks are merged, and the files moved up a level, since the need for "iOS" and "macOS" subdirectories is unnecessary.