Set base SDK to iOS to fix Catalyst build (#365)
We use direct Xcode project integration for Purchases in our app PDF Viewer (i.e. no Swift Package Manager or anything). After updating to 3.7.2, building for Mac Catalyst was failing with this error:

> Building for Mac Catalyst, but the linked framework 'Purchases.framework' is building for macOS. You may need to configure 'Purchases.framework' to build for Mac Catalyst or restrict the platforms for which this framework should be linked in the target editor.

It seems that if the SDKROOT (base SDK) is the default value (macOS), then when building for iOS Xcode infers actually the iOS SDK should be used but when building for Mac Catalyst Xcode uses the macOS SDK instead of the Mac Catalyst SDK for some reason.

I have not tested this with any other integration methods: I’m just providing what worked for us. It seems likely this might break regular macOS builds of Purchases, and if this is the case obviously this PR should be discarded. We’ll just have to apply this patch each time we update if this ends up being the case.