Replaced custom DateFormatter with new ISO8601DateFormatter
Fixes #988.

Unfortunately `ISO8601DateFormatter` inherits from `Formatter`, and not `DateFormatter`, so this introduces a new protocol `DateFormatterType` that both types can implement.

Another unfortunate difference with the new type compared to the old implementation is that milliseconds are either optional or mandatory.
For that reason, I added `ISO8601DateFormatter.default` which replicates the same behavior by deferring `date(from: String) -> Date` to one formatter or the other.
I added tests to cover this behavior directly, but it was only thanks to the existing `ArraySlice` tests that I became aware of this requirement.

Additionally, I made `Transaction.init(with:productId:dateFormatter)` not `public` and not `@objc` so it can take a new `DateFormatterType`.
The method didn't need to be `public`, it was just a pre-Swift-migration leftover.

As a bonus I changed some of the tests in `DateFormatter+ExtensionsTests` to use `XCTUnwrap`.