OfferingsManager: fixed erroneous "Could not find SKProduct" log
The current implementation was computing the intersection of products found in the backend and in the App Store.
If everything was set up correctly, this would lead to the intersection of all products being all of them.

Instead what the implementation needed was to find the products not found in the App Store.