Cleaned up `StoreProduct` abstraction
Fixes [sc-9248].
This is a follow up from the discussion in https://github.com/RevenueCat/purchases-ios/pull/800#discussion_r702058756.

This new implementation avoids having an "abstract" class, which isn't a thing in neither Objective-C nor Swift. So no more `fatalError`s.
Instead, a new protocol `StoreProductType` (which is also Swift only, allowing for more flexibility), introduces the common interface used by the 2 types of product.

To expose a valid type to Objective-C, `StoreProduct` works as type-erasure.

Another benefit of this is that the specific implementations (`SK1StoreProduct` and `SK2StoreProduct`) are no longer `@objc`, so I've turned them into `structs`.