OfferingsManager: fixed erroneous "Could not find SKProduct" log (#1075)
The current implementation was computing the intersection of products found in the backend and in the App Store.
If everything was set up correctly, this would lead to the intersection of all products being all of them.

Instead what the implementation needed was to find the products not found in the App Store.

This lead to showing the following error even when everything was set up correctly:
```
WARN: 🍎‼️ Could not find SKProduct for ["...."] 
There is a problem with your configuration in App Store Connect. 
More info here: https://errors.rev.cat/configuring-products
```

This was already fixed in #868, but looks like it got reverted in #953.