Renamed `ProductInfo` to `ProductRequestData`
The last step for #1045 was to remove `ProductInfo`.
I was planning to simply encode `StoreProduct` for our request serialization, but considering that the format and data that we send to the server is very specific, it still makes sense to have a separate type to represent that.

For that reason, I opted to simply rename the type to a hopefully more clear `ProductRequestData`, and documented its purpose.

After all the prior changes, the conversion from `SK1Product1` to `ProductRequestData` is fairly simple and relies on the same extensions used by `StoreProductType`.

Depends on #1107 and #1108.

- Removed superfluous `Locale.rc_currencyCode()`
- Removed `ProductInfoExtractor` and converted to an extension on `ProductRequestData` (don't need that extra type)
- Changed `ProductInfoExtractorTests` into `ProductRequestDataSK1ProductInitializationTests`, since eventually we'll need to add an equivalent for SK2