Avoid logging user IDs
Fixes https://github.com/RevenueCat/purchases-ios/issues/746 and [sc-10311].

After reviewing the conversations around the issue, I think we can all agree that this is the best solution.
We could make this optional for only `DEBUG` builds, but if a user has a debug build of `RevenueCat` in their app in production, with debug logs enabled (either the default handler or their own), these IDs would be leaking.

### Other changes:
- Moved a few strings into Strings file
- Removed unused strings