Obsoletions: fixed incompatible `introduced` and `obsoleted` annotations
For some reason this only lead to a warning in the Objective-C `APITester` when deprecating a method with an `@objc` annotation:
>  Feature cannot be obsoleted in iOS version 1 before it was introduced in version 12.2; attribute ignored

But looking at the auto-generated `RevenueCat-Swift.h` I could see this was an issue for these methods as well. Likely the `obsoleted` annotation was being ignored.

To fix this, now we only have 1 `@available` annotation for each platform, with a matching `introduced` and `obsoleted` version.