`Purchases.presentCodeRedemptionSheet`: disable on Catalyst
Fixes [sc-10766]. The issue described there is supposed to be fixed in Xcode 13.2. This seems to be isolated to this particular method.
Even though the docs in [`SKPaymentQueue.presentCodeRedemptionSheet`](https://developer.apple.com/documentation/storekit/skpaymentqueue/3566726-presentcoderedemptionsheet) say that it's available on Catalyst 14.0, there is a note:

> This function doesn’t affect Mac apps built with Mac Catalyst.

It crashes when called both from Catalyst and also when running as "Designed for iPad".

To address that, this PR has 2 fixes:
- Disable API for Catalyst: this makes it impossible to call the method on "regular" Catalyst apps
- Disable call at runtime on "Designed for iPad" apps: because they're still "iOS" apps, it's still possible to call the method, but at runtime a new error is logged instead of crashing