Removed unnecessary `SubscriptionPeriod.Unit.unknown`
Follow up to #1220.
In the same way that we obsoleted `StoreProductDiscount.PaymentMode.none`, this follows the same pattern of failing to initialize the `StoreProductDiscount` if the unit is not recognized, instead of exposing an invalid discount.

The type is new in V4, so I think it's safe to not have this in the `V4_API_Migration_guide`, and no obsoletion needed.