`Obsoletions`: added all missing types inside `Purchases.` namespace that were renamed or removed
Fixes [sc-13020].
Currently we only had fix-its for Objective-C names, but in Version 3 the Swift names were namespaced with `Purchases`. Therefore, this adds obsoletions to all types inside `Purchases`.

Note that a few are missing that would conflict with the new types when used inside the `Purchases` class. For example, if we added `Offerings` here, all references to `Offerings` inside of Purchases` would find this new obsoleted type instead of the valid type in the global namespace. The way to address that would be to reference these types with the fully-qualified `RevenueCat.Offerings`, but after consulting with the team everyone agreed that it was not worth the pain moving forward, and therefore some of these are only documented in the migration guide. As for the rest, included here, they will improve the migration experience.