Fixed `ErrorCode`'s domain in Swift
In version 3, `Objective-C` had the constant `RCPurchasesErrorDomain` (`Purchases.ErrorDomain` in Swift), manually defined in Objective-C.
This was also visible to Swift through the constant of the same name, and both had the value `RCPurchasesErrorDomain`.

In version 4, written in Swift, it's not possible to expose global constants to Objective-C. The way the Swift compiler exposes error domains is through the definition of the `Error` itself.
In our case, it's `ErrorCode`, which has an `@objc` name `RCPurchasesErrorCode`. The compiler appends the word "Domain" to that, creating the new `RCPurchasesErrorCodeDomain` which superseded `RCPurchasesErrorDomain`.

However, prior to this change, our implementation of `CustomNSError` was not overriding `errorDomain`, so the value of these constants was the default `RevenueCat.ErrorCode` which was a change (regression) from version 3.0.
Unfortunately because of a bug (https://bugs.swift.org/browse/SR-15841), this isn't enough to fix the domain when `ErrorCode` is converted to `NSError`. This PR includes a workaround, but unfortunately the workaround only provides the correct value to Swift.
`RCPurchasesErrorCodeDomain` will still be the "invalid" `RevenueCat.ErrorCode`, but this is better than nothing.

At least now we have an explicit domain value for Swift errors, and we're exposing a constant for them (`Purchases.ErrorDomain` which was already available in version 3, so it's back now).