`HTTPClient`: started refactor by encapsulating `Request` data
For [sc-9521] and #695.

The new (`private` to `HTTPClient`) type encapsulates all necessary data to make requests. It avoids hardcoding strings like `"GET"`, and it enforces that `POST` requests have a body.
This removes the old `HTTPRequest`, which relied on the old Objective-C `NSCopying` to copy itself, instead of `Swift` `struct`s.

By using this, we avoid passing the same duplicate parameters everywhere, and the same type is used for the `queuedRequests`.