`BackendTests`: fixed `MockHTTPClient` thread safety and SnapshotTesting usage (#1352)
This was a regression that came with the move to `Operation` subclasses. Prior to that, `HTTPClient` was only used from the main thread.

This meant that if the snapshot tests failed, the test runner would crash because `XCTest` doesn't allow adding attachments from outside the main thread.
Additionally, `MockHTTPClient` wasn't thread safe, so without this, it could have lead to race conditions.