Changed `SystemInfo.isApplicationBackgrounded` to use `OperationDispatcher`
This solves a race condition that could cause operations to run out of order (discovered in #1441).

`Purchases.configure` triggers a `getCustomerInfo` call, which relies on `isApplicationBackgrounded`.
The previous implementation always had a runloop "jump", meaning that sometimes a subsequent request might happen out of order.
By using `OperationDispatcher.dispatchOnMainThread`, the jump is avoided if we're already in the main thread, meaning we run everything deterministically in order.