Created `SubscriberAttributesManagerIntegrationTests` (#1441)
I made a lot of changes to how subscriber attributes are synced and the error handling, but we had no test coverage for it other than unit tests. None of the `SubscriberAttributesManager` have completion blocks so we never check if calling `setEmail` etc. actually works in production AFAIK.

This adds coverage for that.