`PurchasesOrchestrator`: fix behavior dealing with `nil` `SKPaymentTransaction.productIdentifier` during purchase
Fixes [CSDK-119].

`PurchasesOrchestrator` uses product identifiers for callback keys.
Previously, `extractProductIdentifier` was allowing `SKPayment.productIdentifier` to be `nil`, as long as the product had an identifier.

However, as seen in the newly added test, this resulted in a purchase call that never invoked its callback, because when the transaction would come back, it wouldn't know which callback to use.

Because of this, there is no point starting the purchase if we can't retrieve the product identifier from the transaction, so now we fail early with an error instead of hanging forever.