Integration Tests: fixed subscription expiration to fix flaky tests
`SKTestSession.expireSubscription` doesn't seem to always work.
To work around it, this manually waits until `expirationDate`.

I've added a warning to make it clear what's happening:
```
INFO: ℹ️ Expiring subscription for product 'com.revenuecat.weekly_1.99.3_day_intro'
WARN: ⚠️ Sleeping for 4.0 seconds to force expiration
```

I've changed all the tests that rely on this to use a weekly subscription. and manually set the `SKTestSession.timeRate` to the lowest value to minimize this duration.