Integration Tests: ensure that environment is sandbox
For some reason the url is `[app_bundle]/StoreKit/receipt` in integration tests.
This is not correct, so I've added a couple of tests: one for entitlements, and another one to check the raw output value of that property.

This is necessary to make sure that receipts are loaded correctly after #1726.