`DeviceCache`: no longer set cache timestamp before beginning request (#1839)
Follow up to [CSDK-391]. As pointed out by @aboedo, we have now 2 caching mechanisms: `CallbackCache` and `DeviceCache` timestamps.
These 2 calls to set the timestamp (through `InMemoryCachedObject.updateCacheTimestamp(date:)`) _before_ starting requests was there before `CallbackCache` was a thing.

However, with that solution, especially after the fix in #1827, the timestamp change is unnecessary. Its purpose was to make sure that a subsequent request didn't trigger an update if it was already in progress, but again, that's already handled by `CallbackCache` now.

The benefit of removing this is also that if a request fails, the timestamp cache wouldn't be an invalid `Date`, since nothing was actually fetched. In practice this isn't an issue because we clear the date on failure (see `OfferingsManager.handleOfferingsUpdateError` for example), but now we wouldn't need to rely on that.

No tests fail after removing this, as expected, since the caching behavior is covered thanks to `CallbackCache`.

_Note: I did check and we can't remove these methods because they are still used elsewhere._

[CSDK-391]: https://revenuecats.atlassian.net/browse/CSDK-391?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ