Added `beginRefundRequest` overload with completion block
Fixes [CSDK-218].

The underlying implementation provided by Apple is `async` only, which is still available to Objective-C. However, to be able to use it from Swift, it must be used with `await`.
This exposes an overload that allows using it with a completion-block based API. Because we don't need these in Objective-C (those are already available), these blocks can use the Swift-only `Result`.