`PaymentQueueWrapper`: handle promotional purchase requests from App Store when SK1 is disabled
Follow up to #1882. After that PR, `StoreKit1Wrapper` was no longer listening to `SKPaymentQueueDelegate.paymentQueue(:shouldAddStorePayment:for:)`, which means that purchases initiated from the App Store were being ignored by the SDK if SK2 is enabled.

To fix this, `PaymentQueueWrapper` now becomes the `SKPaymentQueueDelegate` only when `StoreKit1Wrapper` is not initialized, and `PurchasesOrchestrator` initiates an SK2 purchase instead.

- `PurchasesOrchestrator` purchase methods no longer need the whole `PromotionalOffer`, only the `SignedData`. This allows the new implementation to create that data from only the `SKPaymentDiscount` provided from the `SKPayment`.

- [ ] Test that `StoreKit1Wrapper` is the delegate if SK1 is enabled
- [ ] Test that `PaymentQueueWrapper` is the delegate otherwise
- [ ] Test `PromotionalOffer.SignedData` initialization from SK1
- [ ] Test conversion of SK1 product to SK2.