`StoreKitUnitTests`:  compile on iOS 11.0+
The entire target had a deployment target of iOS 14.0 because `SKTestSession` requires it. However, not all tests in that target use `StoreKitConfigTestCase`, and those weren't running before iOS 14.0.

To fix this, I've removed the deployment target overrides from all targets (expect for integration tests, those have a higher one for simplicity), and instead defined them only once at the *project level*, so all targets inherit that automatically.

`StoreKitConfigTestCase` will skip the test automatically if the iOS is below 14.0, but the other tests will now run on iOS 12 and 13.