Created `CachingProductsManager` to provide consistent caching logic when fetching products
The main feature that this change introduces is providing a _request_ cache to `ProductsFetcherSK2` (like `ProductsFetcherSK` has).
It already implements a _product_ cache, but concurrent requests for the same products result in fetching those multiple times. It does not cache products until they're done being fetched.

Instead of implementing that logic _only_ on `ProductsFetcherSK2` or in `ProductsManager`, this aims to keep the implementation simple in `ProductsManager` and each of the fetchers, and instead uses the decorator pattern to provide that new functionality in a way that's simple to test and maintain, independent of the logic for actually fetching products.

To accomplish that, this adds a new `protocol`: `ProductsManagerType`, with 2 main methods (more about why the need for an SK2 special method below):
```swift
/// Fetches the ``StoreProduct``s with the given identifiers
/// The returned products will be SK1 or SK2 backed depending on the implementation and configuration.
func products(withIdentifiers identifiers: Set<String>, completion: @escaping Completion)

/// Fetches the `SK2StoreProduct`s with the given identifiers.
@available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *)
func sk2Products(withIdentifiers identifiers: Set<String>, completion: @escaping SK2Completion)
```

Those are only completion-block APIs. To avoid repetition, an `extension` on the `protocol` itself provides an `async` API, wrapping on top of these 2 methods, regardless of which `ProductsManagerType` they're being used on.

The hierarchy is now as follows:
- `CachingProductsManager`
    - `ProductsManager`
        - `ProductsFetcherSK1`
        - `ProductsFetcherSK2`

An upcoming PR will get rid of the now duplicate caching logic in `ProductsManager` and the fetchers.

What makes the implementation a bit more convoluted is the need for a special SK2 method.
This is necessary specifically for `TrialOrIntroPriceEligibilityChecker` which requires fetching only SK2 products regardless of `StoreKit2Setting`. In order to guarantee this invariant at compile time through types, a separate cache is maintained with `SK2StoreProduct`s. However, the main caching logic is *not* duplicated and instead shared for both.

The main tests for `CachingProductsManager` are simple unit tests that don't make use of `StoreKitConfigTestCase`.
Because creating SK2 products does require that, I added a separate test class for those.