Extracted `PurchasesType` and `PurchasesSwiftType` (#1912)
This is needed to fix
https://github.com/RevenueCat/purchases-hybrid-common/pull/229


[`purchases-hybrid-common`](https://github.com/RevenueCat/purchases-hybrid-common)
had a [hacky implementation of
`MockPurchases`](https://github.com/RevenueCat/purchases-hybrid-common/blob/4.2.1/ios/PurchasesHybridCommon/PurchasesHybridCommonTests/Mocks/MockPurchases.swift)
that subclassed `Purchases`.
Because of that, it had 100+ lines of code to initialize the instance,
which was never used because all (most) methods were overridden.

This allows a much better way of doing that, **which also enables any
SDK user** to mock `Purchases`.

Another benefit of this is that we can see at a glance what the public
API is.
Also, all docstrings are now extracted into the protocol, which makes
the implementation file shorter. The docs get inherited by the type
itself.

All our API testers pass with no changes, and I've also added the
updated `RevenueCat-Swift.h` to verify that no public API changed.