`BasePurchasesTests`: changed default back to SK1 (#1935)
Follow up to #1922. Most of the tests from the subclasses assume SK2 is
disabled. This changes the default for those back to SK1.
The tests that do require SK2, define so explicitly.

I also replaced a bunch of force-unwraps with `XCTUnwrap` so tests don't
crash (which is what was happening when running these tests with SK2
enabled).