`Integration Tests`: simplified `testIneligibleForIntroAfterPurchaseExpires` to fix flakiness (#1952)
Fixes [CSDK-479].

### Changes:
- Removed extra `verifyEntitlementWentThrough`. #1880 introduced a
change so that weekly subscriptions aren't verified, because they expire
within a second. This make the test flaky due to that race condition.
- Removed `assertNoActiveSubscription`: I tried calling
`SKTestSession.disableAutoRenewForTransaction`, but sometimes the server
still thinks the subscription has auto-renewed. This was making the test
flaky. Turns out that even if it's not active, the eligibility test
still passes as expected.
- Removed call to `syncPurchases`. It doesn't matter what state the
server is in, as long as locally `SKTestSession` has an expired
subscription.

This, together with #1945, should fix the last of the issues causing
flaky integration tests 🤞🏻

[CSDK-479]:
https://revenuecats.atlassian.net/browse/CSDK-479?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ