`OfferingsManager`: added ability to fail if any product is not found
The default behavior is unchanged: if any products aren't found, those are logged, but an `Offerings` value is returned with the found products.

This introduces a new `OfferingsManager.FetchPolicy` that allows receiving an error if any product is not found. For the time being, this is `internal` only, and will be used for [CSDK-451], since we don't want to ignore those errors when using this new SDK tester.