`TrialOrIntroPriceEligibilityChecker`: only use SK2 implementation if enabled
This used to always default to the SK2 implementation if it was available because it was much more reliable.
However, as we've learned (#1893), it can be significantly slow.

Until we implement caching for it ([CSDK-493]), this for now allows users to disable it if it's too slow.

For users that don't specify `useSk2IfAvailable`, SK2 is now the new default, so this class will continue to use the "better" implementation for them too.